fixed issues
I had to remove one of the typedef's in node.h  to compile. It compiles and runs for a little while. The problem is there is a segfault like 3 cycles in. For some reason the number of children that root has is never incremented. To get it to run past 1 loop I had to comment out my exit if statement so that it did not immediately exit after one loop. There may be something not quite right with deconstruct tree also because it gave a message about freeing memory. I see debug statements from node.c saying that there were children nodes created during expansion. I am not sure what is going on. It is either an issue in my code or node.c